Examples: Change the examples to run as individual tests
Please Note: There were a LOT of commented out code and tests in
this repo. These have NOT been migrated to the new test structure
at this time.
