+++ .travis.yml
@@ -22,6 +22,7 @@ install:
     - export PATH=$HOME/.pulumi/bin:$PATH
 before_script:
     - ${PULUMI_SCRIPTS}/ci/ensure-dependencies
+    - echo "$(getent hosts $ARM_LOCATION.management.azure.com | awk '{ print $1 }') management.azure.com" | sudo tee --append /etc/hosts
 script:
     - make travis_${TRAVIS_EVENT_TYPE}
 after_failure:
